Added tests for deep slda and reorganized training epoch to meet current version
